+++ folsom/src/main/java/com/spotify/folsom/client/DefaultRawMemcacheClient.java
@@ -30,6 +30,7 @@ import com.spotify.folsom.Metrics;
 import com.spotify.folsom.RawMemcacheClient;
 import com.spotify.folsom.client.ascii.AsciiMemcacheDecoder;
 import com.spotify.folsom.client.binary.BinaryMemcacheDecoder;
+import com.spotify.folsom.client.tls.SSLEngineFactory;
 import com.spotify.folsom.guava.HostAndPort;
 import com.spotify.folsom.ketama.AddressAndClient;
 import com.spotify.futures.CompletableFutures;
@@ -43,6 +44,7 @@ import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandler;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelOption;
+import io.netty.channel.ChannelPipeline;
 import io.netty.channel.ChannelPromise;
 import io.netty.channel.DefaultChannelPromise;
 import io.netty.channel.EventLoopGroup;
@@ -52,6 +54,7 @@ import io.netty.channel.epoll.EpollSocketChannel;
 import io.netty.channel.nio.NioEventLoopGroup;
 import io.netty.channel.socket.nio.NioSocketChannel;
 import io.netty.handler.codec.DecoderException;
+import io.netty.handler.ssl.SslHandler;
 import io.netty.util.concurrent.DefaultThreadFactory;
 import java.io.IOException;
 import java.net.ConnectException;
@@ -66,6 +69,7 @@ import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Stream;
+import javax.net.ssl.SSLEngine;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -112,7 +116,8 @@ public class DefaultRawMemcacheClient extends AbstractRawMemcacheClient {
       final Metrics metrics,
       final int maxSetLength,
       final EventLoopGroup eventLoopGroup,
-      final Class<? extends Channel> channelClass) {
+      final Class<? extends Channel> channelClass,
+      final SSLEngineFactory sslEngineFactory) {
 
     final ChannelInboundHandler decoder;
     if (binary) {
@@ -124,14 +129,21 @@ public class DefaultRawMemcacheClient extends AbstractRawMemcacheClient {
     final ChannelHandler initializer =
         new ChannelInitializer<Channel>() {
           @Override
-          protected void initChannel(final Channel ch) throws Exception {
-            ch.pipeline()
-                .addLast(
-                    new TcpTuningHandler(),
-                    decoder,
+          protected void initChannel(final Channel ch) {
+            final ChannelPipeline channelPipeline = ch.pipeline();
+            channelPipeline.addLast(new TcpTuningHandler());
 
-                    // Downstream
-                    new MemcacheEncoder());
+            if (sslEngineFactory != null) {
+              final SSLEngine sslEngine =
+                  sslEngineFactory.createSSLEngine(address.getHostText(), address.getPort());
+              SslHandler sslHandler = new SslHandler(sslEngine);
+              // Disable SSL data aggregation
+              // it doesn't play well with memcached protocol and causes connection hangs
+              sslHandler.setWrapDataSize(0);
+              channelPipeline.addLast(sslHandler);
+            }
+
+            channelPipeline.addLast(decoder, new MemcacheEncoder());
           }
         };
 
@@ -236,7 +248,9 @@ public class DefaultRawMemcacheClient extends AbstractRawMemcacheClient {
       // to get better performance in the happy case.
       String disconnectReason = this.disconnectReason.get();
       if (disconnectReason != null) {
-        MemcacheClosedException exception = new MemcacheClosedException(disconnectReason);
+        MemcacheClosedException exception =
+            new MemcacheClosedException(
+                String.format("%s, memcached=%s", disconnectReason, address.getHostText()));
         return onExecutor(CompletableFutures.exceptionallyCompletedFuture(exception));
       }
 
@@ -456,6 +470,10 @@ public class DefaultRawMemcacheClient extends AbstractRawMemcacheClient {
       // Use the pending counter as a way of marking disconnected for performance reasons
       // Once we are disconnected we will not really decrease this value any more anyway.
       pendingCounter.set(pendingCounterLimit);
+      SslHandler sslHandler = channel.pipeline().get(SslHandler.class);
+      if (sslHandler != null) {
+        sslHandler.closeOutbound();
+      }
       channel.close();
       GLOBAL_CONNECTION_COUNT.decrementAndGet();
       metrics.unregisterOutstandingRequestsGauge(pendingRequestGauge);