+++ folsom/src/main/java/com/spotify/folsom/client/ascii/AsciiResponse.java
@@ -27,6 +27,7 @@ public class AsciiResponse {
   public static final AsciiResponse ERROR = new AsciiResponse(Type.ERROR);
   public static final AsciiResponse VALUE_TOO_LARGE = new AsciiResponse(Type.VALUE_TOO_LARGE);
   public static final AsciiResponse OUT_OF_MEMORY = new AsciiResponse(Type.OUT_OF_MEMORY);
+  public static final AsciiResponse CLIENT_ERROR = new AsciiResponse(Type.CLIENT_ERROR);
 
   public final Type type;
 
@@ -48,7 +49,8 @@ public class AsciiResponse {
     EMPTY_LIST,
     ERROR,
     VALUE_TOO_LARGE,
-    OUT_OF_MEMORY
+    OUT_OF_MEMORY,
+    CLIENT_ERROR
   }
 
   @Override